refactor(cli): extract `mapa` viz command into cli_viz.py (2b)
The ~400-line `mapa` command (Plotly HTML semantic-map generator) was inline
in cli.py. Moved verbatim to cli_viz.py — it only referenced Config + console,
so the extraction is mechanical. cli.py 2710 → 2312 LOC.
